Fix compilation warning
fmtlib bundled with spdlog doesn't like direct printing of
enums. Let's just cast.

Reviewed-on: https://codereview.kdab.com/c/kdab/kdutils/+/136319
Tested-by: Continuous Integration <build@kdab.com>

+++ examples/gui_window/gui_window.cpp
@@ -51,12 +51,12 @@ class ExampleWindow : public KDGui::Window
 
     void keyPressEvent(KDGui::KeyPressEvent *ev) override
     {
-        SPDLOG_INFO("{}() key = {}", __FUNCTION__, ev->key());
+        SPDLOG_INFO("{}() key = {}", __FUNCTION__, static_cast<int>(ev->key()));
     }
 
     void keyReleaseEvent(KDGui::KeyReleaseEvent *ev) override
     {
-        SPDLOG_INFO("{}() key = {}", __FUNCTION__, ev->key());
+        SPDLOG_INFO("{}() key = {}", __FUNCTION__, static_cast<int>(ev->key()));
     }
 };
